feat: move admin appointments to another file
